Moves classes in different files.
