+++ src/HtmlLabel/Android/Renderer.android.cs
@@ -9,7 +9,6 @@ using Android.Widget;
 using Java.Lang;
 using LabelHtml.Forms.Plugin.Abstractions;
 using LabelHtml.Forms.Plugin.Droid;
-using Org.Xml.Sax;
 using Xamarin.Essentials;
 using Xamarin.Forms;
 using Xamarin.Forms.Internals;
@@ -18,10 +17,10 @@ using Xamarin.Forms.Platform.Android;
 [assembly: ExportRenderer(typeof(HtmlLabel), typeof(HtmlLabelRenderer))]
 namespace LabelHtml.Forms.Plugin.Droid
 {
-    /// <summary>
-    /// HtmlLable Implementation
-    /// </summary>
-    [Preserve(AllMembers = true)]
+	/// <summary>
+	/// HtmlLable Implementation
+	/// </summary>
+	[Preserve(AllMembers = true)]
     public class HtmlLabelRenderer : LabelRenderer
     {
 		private const string _tagUlRegex = "[uU][lL]";
@@ -182,183 +181,6 @@ namespace LabelHtml.Forms.Plugin.Droid
 			}
 
 			return builder;
-		}
-	}
-
-	// TagHandler that handles lists (ul, ol)
-	internal class ListTagHandler : Java.Lang.Object, Html.ITagHandler
-	{
-		public const string TagUlc = "ULC";
-		public const string TagOlc = "OLC";
-		public const string TagLic = "LIC";
-
-		private ListBuilder _listBuilder = new ListBuilder();
-		
-		public void HandleTag(bool opening, string tag, IEditable output, IXMLReader xmlReader)
-		{
-			tag = tag.ToUpperInvariant();
-			if (tag.Equals(TagLic, StringComparison.Ordinal))
-			{
-				_listBuilder.Li(opening, output);
-				return;
-			}
-			if (opening)
-			{
-				if (tag.Equals(TagOlc, StringComparison.Ordinal))
-				{
-					_listBuilder = _listBuilder.StartList(true, output);
-					return;
-				}
-				if (tag.Equals(TagUlc, StringComparison.Ordinal))
-				{
-					_listBuilder = _listBuilder.StartList(false, output);
-					return;
-				}
-			}
-			else
-			{
-				if (tag.Equals(TagOlc, StringComparison.Ordinal) || tag.Equals(TagUlc, StringComparison.Ordinal))
-				{
-					_listBuilder = _listBuilder.CloseList(output);
-				}
-				return;
-			}
-		}
-	}
-
-	internal class ListBuilder
-	{
-		private const int _listIndent = 20;
-
-		private readonly int _gap = 0;
-		private readonly LiGap _liGap;
-		private readonly ListBuilder _parent = null;
-
-		private int _liIndex = -1;
-		private int _liStart = -1;
-		
-		internal ListBuilder() : this(null)
-		{
-		}
-
-		internal ListBuilder(LiGap liGap)
-		{
-			_parent = null;
-			_gap = 0;
-			_liGap = liGap ?? GetLiGap(null);
-		}
-
-		private ListBuilder(ListBuilder parent, bool ordered)
-		{
-			_parent = parent;
-			_liGap = parent._liGap;
-			_gap = parent._gap + _listIndent + _liGap.GetGap(ordered);
-			_liIndex = ordered ? 0 : -1;
-		}
-
-		internal ListBuilder StartList(bool ordered, IEditable output)
-		{
-			if (_parent == null)
-			{
-				if (output.Length() > 0)
-				{
-					_ = output.Append("\n ");
-				}
-			}
-			return new ListBuilder(this, ordered);
-		}
-
-		private bool IsOrdered()
-		{
-			return _liIndex >= 0;
-		}
-
-		internal void Li(bool opening, IEditable output)
-		{
-			if (opening)
-			{
-				EnsureParagraphBoundary(output);
-				_liStart = output.Length();
-
-				_ = IsOrdered() ? output.Append(++_liIndex + ". ") : output.Append("•  ");
-			}
-			else
-			{
-				if (_liStart >= 0)
-				{
-					EnsureParagraphBoundary(output);
-					using var leadingMarginSpan = new LeadingMarginSpanStandard(_gap - _liGap.GetGap(IsOrdered()), _gap);
-					output.SetSpan(leadingMarginSpan, _liStart, output.Length(), SpanTypes.ExclusiveExclusive);
-					_liStart = -1;
-				}
-			}
-		}
-
-
-		internal ListBuilder CloseList(IEditable output)
-		{
-			EnsureParagraphBoundary(output);
-			ListBuilder result = _parent;
-			if (result == null)
-			{
-				result = this;
-			}
-
-			if (result._parent == null)
-			{
-				_ = output.Append('\n');
-			}
-
-			return result;
-		}
-
-		private static void EnsureParagraphBoundary(IEditable output)
-		{
-			if (output.Length() == 0)
-			{
-				return;
-			}
-
-			var lastChar = output.CharAt(output.Length() - 1);
-			if (lastChar != '\n')
-			{
-				_ = output.Append('\n');
-			}
-		}
-
-		internal class LiGap
-		{
-			private readonly int _orderedGap;
-			private readonly int _unorderedGap;
-
-			internal LiGap(int orderedGap, int unorderedGap)
-			{
-				_orderedGap = orderedGap;
-				_unorderedGap = unorderedGap;
-			}
-
-			public int GetGap(bool ordered)
-			{
-				return ordered ? _orderedGap : _unorderedGap;
-			}
-		}
-
-		internal static LiGap GetLiGap(TextView tv)
-		{
-			return tv == null ? 
-				new LiGap(40, 30) : 
-				new LiGap(ComputeWidth(tv, true), ComputeWidth(tv, false));
-		}
-
-		private static int ComputeWidth(TextView tv, bool ordered)
-		{
-			Android.Graphics.Paint paint = tv.Paint;
-			using var bounds = new Android.Graphics.Rect();
-			var myString = ordered ? "99. " : "• ";
-		    paint.GetTextBounds(myString, 0, myString.Length, bounds);
-			var width = bounds.Width();
-			var pt = Android.Util.TypedValue.ApplyDimension(Android.Util.ComplexUnitType.Pt, width, tv.Context.Resources.DisplayMetrics);			
-			return (int)pt;
 		}
 	}
 }
